requirements.txt
# Core dependencies
anthropic>=0.40.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0

# Analysis and visualization
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0

# Jupyter notebooks (optional)
jupyter>=1.0.0
ipykernel>=6.0.0

# Development (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
mypy>=1.5.0
